+++ src/components/bootstrap/bootstrap.tsx
@@ -2,8 +2,8 @@ import '../../../node_modules/bootstrap/dist/css/bootstrap.css';
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { Credentials, CredentialsInterface } from '../../services/credentials';
-import { GithubApi } from '../../services/github-api';
-import { getUserDetails } from '../../stores/github-store';
+import { GithubApi, GithubRepoInterface, GithubIssuesInterface} from '../../services/github-api';
+import { getUserDetails, getUserSubscriptions , getUserRepositories, getIssuesForUserRepository, getIssuesForUserSubscription } from '../../stores/github-store';
 import Footer from '../footer/footer';
 import Header from '../header/header';
 import Navigation from '../navigation/navigation';
@@ -21,6 +21,8 @@ class Bootstrap extends React.Component<any, any> {
     this.githubApi = new GithubApi(this.credentials);
 
     this.getUserDetails();
+    this.getUserRepositoriesWithIssues();
+    this.getUserSubscriptionsWithIssues();
   }
 
   private getUserDetails() {
@@ -30,6 +32,49 @@ class Bootstrap extends React.Component<any, any> {
       dispatch(getUserDetails(response));
     });
   }
+
+  private getUserRepositoriesWithIssues(nextReposUrl?: string) {
+    let dispatch = this.props.dispatch;
+    let url: string = nextReposUrl ? nextReposUrl : null;
+
+    this.githubApi.getUserRepositories(url).then((response: any) => {
+      dispatch(getUserRepositories(response));
+
+      //TODO move offsetIdx logic into a central place
+      response.repos.map((repo: GithubRepoInterface, index: number) => {
+        let offsetIdx = nextReposUrl ? (this.state.repositories.length - 30) + index : index;
+
+        this.githubApi.getIssuesForRepository(repo.details.name, repo.details.owner.login).then((response: GithubIssuesInterface) => {
+          dispatch(getIssuesForUserRepository(response, offsetIdx));
+        })
+      })
+    });
+  }
+
+  private getUserSubscriptionsWithIssues(nextReposUrl?: string) {
+    console.log('getUserSubscriptionsWithIssues state', this.state);
+    console.log('getUserSubscriptionsWithIssues props', this.props);
+    let dispatch = this.props.dispatch;
+    let url: string = nextReposUrl ? nextReposUrl : null;
+    this.githubApi.getUserSubscriptions(url).then((response: any) => {
+      console.log('getUserSubscriptionsWithIssues response', response)
+      dispatch(getUserSubscriptions(response));
+
+      //TODO move offsetIdx logic into a central place
+      // response.repos.forEach((repo: GithubRepoInterface, index: number) => {
+      //   console.log('index', index);
+      //   let offsetIdx = nextReposUrl ? ((this.state.repositories.length + index) - 30) : index;
+      //   console.log('length', this.state.repositories.length);
+      //   //console.log('division is hard', this.state.repositories.length / 30);
+      //   console.log('offsetIdx', offsetIdx);
+      //
+      //   this.githubApi.getIssuesForRepository(repo.details.name, repo.details.owner.login).then((response: GithubIssuesInterface) => {
+      //     dispatch(getIssuesForUserSubscription(response, offsetIdx));
+      //   })
+      // })
+    });
+  }
+
 
   render() {
     return (
