+++ src/services/github-api.ts
@@ -107,10 +107,9 @@ export class GithubApi {
     });
   }
 
-  getUserRepositories (username?: string, nextUrl?: string): any {
+  getUserRepositories (nextUrl?: string): any {
     // TODO should this even be required since its a call specifically for the user?
-    let user = username || this.credentials.username;
-    let url = nextUrl || this.baseUrl + 'users/' + user + '/repos';
+    let url = nextUrl || this.baseUrl + 'users/' + this.credentials.username + '/repos';
 
     return axios.get(url).then((response: any) => {
       let modeledRepos: Array<GithubRepoInterface> = [];
@@ -134,10 +133,9 @@ export class GithubApi {
     });
   }
 
-  getUserSubscriptions (username?: string, nextUrl?: string): any {
+  getUserSubscriptions (nextUrl?: string): any {
     // TODO should this even be required since its a call specifically for the user?
-    let user = username || this.credentials.username;
-    let url = nextUrl || this.baseUrl + 'users/' + user + '/subscriptions';
+    let url = nextUrl || this.baseUrl + 'users/' + this.credentials.username + '/subscriptions';
 
     return axios.get(url).then((response: any) => {
       let modeledRepos: Array<GithubRepoInterface> = [];
