+++ src/views/personal/personal.tsx
@@ -1,9 +1,6 @@
 import * as React from 'react';
 import RepositoriesTable from '../../components/repositories-table/repositories-table';
 import { connect } from 'react-redux';
-import { Credentials, CredentialsInterface } from '../../services/credentials';
-import { GithubApi, GithubIssuesInterface, GithubRepoInterface } from '../../services/github-api';
-import { getUserRepositories, getIssuesForUserRepository } from '../../stores/github-store';
 
 function mapStateToProps(state) {
   return {
@@ -15,38 +12,15 @@ function mapStateToProps(state) {
 
 //TODO change use any, any to use types
 class Personal extends React.Component<any, any>{
-  private credentials: CredentialsInterface;
-  private githubApi: any;
 
   constructor(props) {
     super(props);
 
-    this.credentials = new Credentials().getCredentials();
-    this.githubApi = new GithubApi(this.credentials);
     this.state = {
       repositories: [],
       hasMoreRepos: false,
       nextReposUrl: ''
     };
-
-    this.getUserRepositoriesWithIssues();
-  }
-
-  private getUserRepositoriesWithIssues(nextReposUrl?: string) {
-    let dispatch = this.props.dispatch;
-
-    this.githubApi.getUserRepositories(null, nextReposUrl).then((response: any) => {
-      dispatch(getUserRepositories(response));
-
-      //TODO move offsetIdx logic into a central place
-      response.repos.map((repo: GithubRepoInterface, index: number) => {
-        let offsetIdx = nextReposUrl ? (this.state.repositories.length - 30) + index : index;
-
-        this.githubApi.getIssuesForRepository(repo.details.name, repo.details.owner.login).then((response: GithubIssuesInterface) => {
-          dispatch(getIssuesForUserRepository(response, offsetIdx));
-        })
-      })
-    });
   }
 
   componentWillReceiveProps(nextProps) {
@@ -59,13 +33,18 @@ class Personal extends React.Component<any, any>{
 
   //this is here since if a component isnt mounted when dispatches to the store happen
   //the component will need to manually query the store to hydrate itself into state
-  // componentWillMount () {
-  //   this.setState({
-  //     repositories: this.props.repositories,
-  //     hasMoreRepos: this.props.hasMoreRepos,
-  //     nextReposUrl: this.props.nextReposUrl
-  //   });
-  // }
+  componentWillMount () {
+    console.log('component will mount');
+    // this.setState({
+    //   repositories: this.props.repositories,
+    //   hasMoreRepos: this.props.hasMoreRepos,
+    //   nextReposUrl: this.props.nextReposUrl
+    // });
+  }
+
+  private getUserRepositoriesWithIssues(){
+    console.log('TODO load more -> call up');
+  }
 
   render() {
     return (
